requirements.txt
tensorflow>=2.10.0
opencv-python>=4.7.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
pyyaml>=6.0
pytest>=7.0.0
